Moved files outside examples folder
