+++ applications/pvc/flash_attention_v2/collective/pvc_flash_attn_epilogue.hpp
@@ -39,8 +39,6 @@
 #include "cutlass/epilogue/dispatch_policy.hpp"
 #include "cutlass/epilogue/collective/collective_epilogue.hpp"
 #include "cutlass/epilogue/collective/detail.hpp"
-#include "cutlass/epilogue/fusion/callbacks.hpp"
-#include "cutlass/epilogue/fusion/sm90_visitor_tma_warpspecialized.hpp"
 #include "cutlass/detail/layout.hpp"
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
